+++ lib/stellar_spectrum/services/get_sequence_number.rb
@@ -2,13 +2,15 @@ module StellarSpectrum
   class GetSequenceNumber
 
     extend LightService::Action
-    expects :stellar_client, :address
-    promises :sequence_number
+    expects :stellar_client, :channel_account
+    promises :current_sequence_number, :next_sequence_number
 
     executed do |c|
-      account = Stellar::Account.from_address(c.address)
-      account_info = c.stellar_client.account_info(account)
-      c.sequence_number = account_info.sequence.to_i
+      account_info = c.stellar_client.account_info(c.channel_account)
+      c.current_sequence_number = account_info.sequence.to_i
+
+      c.next_sequence_number = c[:force_sequence_number] ||
+        c.current_sequence_number + 1
     end
 
   end