+++ lib/stellar_spectrum/services/unlocking/get_account_to_unlock.rb
@@ -1,0 +1,23 @@
+module StellarSpectrum
+  module Unlocking
+    class GetAccountToUnlock
+
+      extend LightService::Action
+      expects :locked_accounts
+      promises :channel_account
+
+      executed do |c|
+        address, info = c.locked_accounts.except(c[:except_address])
+          .sort_by {|address, info| info[:pttl]}.last
+
+        if address.present?
+          c.channel_account = Stellar::Account.from_address(address)
+        else
+          c.channel_account = nil
+          c.fail_and_return! "No addresses to unlock"
+        end
+      end
+
+    end
+  end
+end