+++ lib/stellar_spectrum/services/unlocking/unlock.rb
@@ -3,11 +3,11 @@ module StellarSpectrum
     class Unlock
 
       extend LightService::Action
-      expects :redis, :address
+      expects :redis, :channel_account
       promises :unlock_response
 
       executed do |c|
-        address_key = GetKeyForAddress.execute(c.address)
+        address_key = GetKeyForAddress.execute(c.channel_account.address)
         c.unlock_response = c.redis.del(address_key)
       end