+++ spec/lib/stellar_spectrum/client_spec.rb
@@ -4,7 +4,8 @@ module StellarSpectrum
   RSpec.describe Client do
 
     let(:stellar_client) do
-      Stellar::Client.new(horizon: CONFIG[:horizon_url])
+      InitStellarClient.execute(horizon_url: CONFIG[:horizon_url])
+        .stellar_client
     end
     let(:from_account) { Stellar::Account.from_seed(CONFIG[:sender_seed]) }
     let(:destination_account) do
@@ -30,7 +31,7 @@ module StellarSpectrum
 
       tx_hash = tx._attributes["hash"]
 
-      tx = client.stellar_client.horizon.transaction(hash: tx_hash)
+      tx = stellar_client.horizon.transaction(hash: tx_hash)
       expect(tx.source_account).to_not eq from_account.address
 
       operation = tx.operations.records.first
@@ -56,7 +57,7 @@ module StellarSpectrum
 
       tx_hash = tx._attributes["hash"]
 
-      tx = client.stellar_client.horizon.transaction(hash: tx_hash)
+      tx = stellar_client.horizon.transaction(hash: tx_hash)
       expect(tx.memo).to eq "MOMO"
     end
 
@@ -98,6 +99,83 @@ module StellarSpectrum
           to eq(5 * 1)
       end
       WebMock.disable_net_connect!
+    end
+
+    context "transaction_source and sequence are given" do
+      let(:channel_account) do
+        Stellar::Account.from_seed(seeds.first)
+      end
+      let(:seeds) do
+        # Return one seed so we can be sure the first time around we send the tx
+        # that we pick the only available seed
+        [CONFIG[:payment_channel_seeds].last]
+      end
+
+      it "uses the given transaction_source and sequence", vcr: {record: :once} do
+        client = described_class.new(
+          redis_url: CONFIG[:redis_url],
+          seeds: CONFIG[:payment_channel_seeds],
+          horizon_url: CONFIG[:horizon_url],
+          seeds: seeds,
+        )
+
+        next_sequence_number = GetSequenceNumber.execute(
+          stellar_client: stellar_client,
+          channel_account: channel_account,
+        ).next_sequence_number
+
+        tx_0 = client.send_payment(
+          from: from_account,
+          to: destination_account,
+          amount: Stellar::Amount.new(1),
+          memo: "MOMO",
+        )
+
+        expect(tx_0._response).to be_success
+        tx_0_hash = tx_0._attributes["hash"]
+
+        tx_1 = client.send_payment(
+          from: from_account,
+          to: destination_account,
+          amount: Stellar::Amount.new(1),
+          memo: "MOMO",
+          transaction_source: channel_account,
+          sequence: next_sequence_number,
+        )
+
+        expect(tx_1._response).to be_success
+        tx_1_hash = tx_1._attributes["hash"]
+
+        expect(tx_0_hash).to eq tx_1_hash
+      end
+    end
+
+    context "sending payment times out" do
+      let(:channel_account) do
+        Stellar::Account.from_seed(seeds.first)
+      end
+
+      it "retries until it is able to send it" do
+        VCR.use_cassette(
+          "StellarSpectrum_Client/"\
+          "sending_payment_times_out/"\
+          "definition-account_info-post_tx_timeout_2x_then_success",
+        ) do
+          client = described_class.new(
+            redis_url: CONFIG[:redis_url],
+            seeds: CONFIG[:payment_channel_seeds],
+            horizon_url: CONFIG[:horizon_url],
+          )
+
+          result = client.send_payment(
+            from: from_account,
+            to: destination_account,
+            amount: Stellar::Amount.new(1),
+          )
+
+          expect(result._success?).to be true
+        end
+      end
     end
 
   end